requirements-dev.txt
# Django and Core Dependencies
Django==4.2.7
psycopg2-binary==2.9.9
django-environ==0.11.2
gunicorn==21.2.0
whitenoise==6.6.0

# API and REST
djangorestframework==3.14.0
django-filter==23.5
django-cors-headers==4.3.1

# Celery and Task Queue
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# AI and External APIs
openai>=1.0.0
PyGithub>=2.0.0
requests>=2.31.0
GitPython>=3.1.0

# Testing Tools
pytest==7.3.1
pytest-django==4.7.0
pytest-cov==4.0.0
pytest-mock<3.10.1
pytest-playwright
coverage
axe-playwright-python
selenium
beautifulsoup4

# Linters
ruff
black==23.3.0
flake8
pylint==2.17.4

# Development Tools
cruft
pip-tools
factory-boy==3.3.0

# Build and Release
build
twine